Solution1

+++ app/main.py
@@ -18,21 +18,20 @@ class CarWashStation:
         self,
         distance_from_city_center: int,
         clean_power: int,
-        average_rating: int,
+        average_rating: float,
         count_of_ratings: int,
     ) -> None:
         self.distance_from_city_center = distance_from_city_center
         self.clean_power = clean_power
         self.average_rating = round(average_rating, 1)
         self.count_of_ratings = count_of_ratings
-        # car = Car
 
     def serve_cars(self, cars: List[Car]) -> float:
         price = 0
         for car in cars:
             if car.clean_mark < self.clean_power:
                 price += self.calculate_washing_price(car)
-                self.wash_single_car(car)
+            self.wash_single_car(car)
         return round(price, 1)
 
     def calculate_washing_price(self, car: Car) -> float:
